requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
browser-use>=0.12.0
langchain-google-genai>=1.0.0
python-dotenv==1.1.0
playwright
